refactor: Move iWork functions in Toolkit back to iWork Class
