+++ lib/Toolkit.js
@@ -1960,565 +1960,15 @@ class Toolkit {
     this.conn.run(toJson); // Post the input XML and get the response.
   }
 
-  getSysValue(sysValue, cb) {
-    let keyValid = false;
-    let type = '10i0';
 
-    sysvalArray.some(value => value.key.some((key) => {
-      if (sysValue === key) {
-        keyValid = true;
-        ({ type } = value);
-        return true;
-      }
-      return false;
-    }));
 
-    // TODO: Throw error
-    if (keyValid === false) { return; }
 
-    let item;
-    if (type === '10i0') { item = { type: '10i0', value: 0 }; } else { item = { type, value: '' }; }
 
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec1',
-      fields: [
-        { type: '10i0', value: 0 }, // Number of system values returned
-        { type: '10i0', value: 0 }, // Offset to system value information table
-        { type: '10A', value: '' }, // System value
-        { type: '1A', value: '' }, // Type of data(C--character / B--binary / blank--not available.)
-        { type: '1A', value: '' }, // Information status(blank--The information was available.)
-        { type: '10i0', value: 0 }, // Length of data
-        item,
-      ],
-    };
 
-    const pgm = new ProgramCall('QWCRSVAL', { lib: 'QSYS' });
-    pgm.addParam(outBuf); // Receiver buffer
-    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 }); // Length of the receiver buffer
-    pgm.addParam({ type: '10i0', value: 1 }); // Number of system values to retrieve
-    pgm.addParam({ type: '10A', value: sysValue }); // System value name
-    pgm.addParam(this.errno); // Error code
 
-    this.conn.add(pgm.toXML());
 
-    let rtValue;
 
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-          rtValue = data[6]._; // Get the returned value from the output array.
-        } else {
-          rtValue = str;
-        }
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
 
-    this.conn.run(toJson); // Post the input XML
-  }
-
-  getSysStatus(cb) {
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec1',
-      fields: [
-        { type: '10i0', value: 0 }, // Bytes available
-        { type: '10i0', value: 0 }, // Bytes returned
-        { type: '20u0', value: '' }, // Current date and time
-        { type: '8A', value: '' }, // System name
-        { type: '10i0', value: 0 }, // Users currently signed on
-        { type: '10i0', value: 0 }, // Users temporarily signed off (disconnected)
-        { type: '10i0', value: 0 }, // Users suspended by system request
-        { type: '10i0', value: 0 }, // Users suspended by group jobs
-        { type: '10i0', value: 0 }, // Users signed off with printer output waiting to print
-        { type: '10i0', value: 0 }, // Batch jobs waiting for messages
-        { type: '10i0', value: 0 }, // Batch jobs running
-        { type: '10i0', value: 0 }, // Batch jobs held while running
-        { type: '10i0', value: 0 }, // Batch jobs ending
-        { type: '10i0', value: 0 }, // Batch jobs waiting to run or already scheduled
-        { type: '10i0', value: 0 }, // Batch jobs held on a job queue
-        { type: '10i0', value: 0 }, // Batch jobs on a held job queue
-        { type: '10i0', value: 0 }, // Batch jobs on an unassigned job queue
-        { type: '10i0', value: 0 }, // Batch jobs ended with printer output waiting to print
-      ],
-    };
-    const pgm = new ProgramCall('QWCRSSTS', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
-    pgm.addParam({ type: '8A', value: 'SSTS0100' });
-    pgm.addParam({ type: '10A', value: '*NO' });
-    pgm.addParam(this.errno);
-
-    this.conn.add(pgm.toXML());
-
-    let rtValue;
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-          rtValue = {
-            Current_date_and_time: data[2]._,
-            System_name: data[3]._,
-            Users_currently_signed_on: data[4]._,
-            'Users_temporarily_signed_off_(disconnected)': data[5]._,
-            Users_suspended_by_system_request: data[6]._,
-            Users_suspended_by_group_jobs: data[7]._,
-            Users_signed_off_with_printer_output_waiting_to_print: data[8]._,
-            Batch_jobs_waiting_for_messages: data[9]._,
-            Batch_jobs_running: data[10]._,
-            Batch_jobs_held_while_running: data[11]._,
-            Batch_jobs_ending: data[12]._,
-            Batch_jobs_waiting_to_run_or_already_scheduled: data[13]._,
-            Batch_jobs_held_on_a_job_queue: data[14]._,
-            Batch_jobs_on_a_held_job_queue: data[15]._,
-            Batch_jobs_on_an_unassigned_job_queue: data[16]._,
-            Batch_jobs_ended_with_printer_output_waiting_to_print: data[17]._,
-          };
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson);
-  }
-
-  getSysStatusExt(cb) {
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec1',
-      fields: [
-        { type: '10i0', value: 0 }, // Bytes available
-        { type: '10i0', value: 0 }, // Bytes returned
-        { type: '20u0', value: '' }, // Current date and time
-        { type: '8A', value: '' }, // System name
-        { type: '6A', value: '' }, // Elapsed time
-        { type: '1A', value: '' }, // Restricted state flag
-        { type: '1A', value: '' }, // Reserved
-        { type: '10i0', value: 0 }, // % processing unit used
-        { type: '10i0', value: 0 }, // Jobs in system
-        { type: '10i0', value: 0 }, // % permanent addresses
-        { type: '10i0', value: 0 }, // % temporary addresses
-        { type: '10i0', value: 0 }, // System ASP
-        { type: '10i0', value: 0 }, // % system ASP used
-        { type: '10i0', value: 0 }, // Total auxiliary storage
-        { type: '10i0', value: 0 }, // Current unprotected storage used
-        { type: '10i0', value: 0 }, // Maximum unprotected storage used
-        { type: '10i0', value: 0 }, // % DB capability
-        { type: '10i0', value: 0 }, // Main storage size
-        { type: '10i0', value: 0 }, // Number of partitions
-        { type: '10i0', value: 0 }, // Partition identifier
-        { type: '10i0', value: 0 }, // Reserved
-        { type: '10i0', value: 0 }, // Current processing capacity
-        { type: '1A', value: '' }, // Processor sharing attribute
-        { type: '3A', value: '' }, // Reserved
-        { type: '10i0', value: 0 }, // Number of processors
-        { type: '10i0', value: 0 }, // Active jobs in system
-        { type: '10i0', value: 0 }, // Active threads in system
-        { type: '10i0', value: 0 }, // Maximum jobs in system
-        { type: '10i0', value: 0 }, // % temporary 256MB segments used
-        { type: '10i0', value: 0 }, // % temporary 4GB segments used
-        { type: '10i0', value: 0 }, // % permanent 256MB segments used
-        { type: '10i0', value: 0 }, // % permanent 4GB segments used
-        { type: '10i0', value: 0 }, // % current interactive performance
-        { type: '10i0', value: 0 }, // % uncapped CPU capacity used
-        { type: '10i0', value: 0 }, // % shared processor pool used
-        { type: '20u0', value: 0 }, // Main storage size (long)
-      ],
-    };
-    const pgm = new ProgramCall('QWCRSSTS', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
-    pgm.addParam({ type: '8A', value: 'SSTS0200' });
-    pgm.addParam({ type: '10A', value: '*NO' });
-    pgm.addParam(this.errno);
-
-    this.conn.add(pgm.toXML());
-
-    let rtValue;
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-          rtValue = {
-            Current_date_and_time: data[2]._,
-            System_name: data[3]._,
-            Elapsed_time: data[4]._,
-            Restricted_state_flag: data[5]._,
-            '%_processing_unit_used': data[7]._,
-            Jobs_in_system: data[8]._,
-            '%_permanent_addresses': data[9]._,
-            '%_temporary_addresses': data[10]._,
-            System_ASP: data[11]._,
-            '%_system_ASP_used': data[12]._,
-            Total_auxiliary_storage: data[13]._,
-            Current_unprotected_storage_used: data[14]._,
-            Maximum_unprotected_storage_used: data[15]._,
-            '%_DB_capability': data[16]._,
-            Main_storage_size: data[17]._,
-            Number_of_partitions: data[18]._,
-            Partition_identifier: data[19]._,
-            Current_processing_capacity: data[21]._,
-            Processor_sharing_attribute: data[22]._,
-            Number_of_processors: data[24]._,
-            Active_jobs_in_system: data[25]._,
-            Active_threads_in_system: data[26]._,
-            Maximum_jobs_in_system: data[27]._,
-            '%_temporary_256MB_segments_used': data[28]._,
-            '%_temporary_4GB_segments_used': data[29]._,
-            '%_permanent_256MB_segments_used': data[30]._,
-            '%_permanent_4GB_segments_used': data[31]._,
-            '%_current_interactive_performance': data[32]._,
-            '%_uncapped_CPU_capacity_used': data[33]._,
-            '%_shared_processor_pool_used': data[34]._,
-            'Main_storage_size_(long)': data[35]._,
-          };
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson);
-  }
-
-  getJobStatus(jobId, cb) {
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec1',
-      fields: [
-        { type: '10i0', value: 0 }, // Bytes returned
-        { type: '10i0', value: 0 }, // Bytes available
-        { type: '10A', value: '' }, // Job status
-        { type: '16h', value: '' }, // Internal job identifier
-        { type: '26A', value: '' }, // Fully qualified job name
-      ],
-    };
-    const pgm = new ProgramCall('QWCRJBST', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
-    pgm.addParam({ type: '6A', value: jobId });
-    pgm.addParam({ type: '8A', value: 'JOBS0100' });
-    pgm.addParam(this.errno);
-
-    this.conn.add(pgm.toXML());
-
-    let rtValue;
-
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-          rtValue = {
-            Job_status: data[2]._,
-            Fully_qualified_job_name: data[4]._,
-          };
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson);
-  }
-
-  getJobInfo(jobName, userName, jobNumber, cb) {
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec1',
-      fields: [
-        { type: '10i0', value: 0 }, // Number of bytes returned
-        { type: '10i0', value: 0 }, // Number of bytes available
-        { type: '10A', value: '' }, // Job name
-        { type: '10A', value: '' }, // User name
-        { type: '6A', value: '' }, // Job number
-        { type: '16h', value: '' }, // Internal job identifier
-        { type: '10A', value: '' }, // Job status
-        { type: '1A', value: '' }, // Job type
-        { type: '1A', value: '' }, // Job subtype
-        { type: '10A', value: '' }, // Subsystem description name
-        { type: '10i0', value: 0 }, // Run priority (job)
-        { type: '10i0', value: 0 }, // System pool identifier
-        { type: '10i0', value: 0 }, // Processing unit time used, if less than 2,147,483,647 milliseconds
-        { type: '10i0', value: 0 }, // Number of auxiliary I/O requests, if less than 2,147,483,647
-        { type: '10i0', value: 0 }, // Number of interactive transactions
-        { type: '10i0', value: 0 }, // Response time total
-        { type: '1A', value: '' }, // Function type
-        { type: '10A', value: '' }, // Function name
-        { type: '4A', value: '' }, // Active job status
-        { type: '10i0', value: 0 }, // Number of database lock waits
-        { type: '10i0', value: 0 }, // Number of internal machine lock waits
-        { type: '10i0', value: 0 }, // Number of nondatabase lock waits
-        { type: '10i0', value: 0 }, // Time spent on database lock waits
-        { type: '10i0', value: 0 }, // Time spent on internal machine lock waits
-        { type: '10i0', value: 0 }, // Time spent on nondatabase lock waits
-        { type: '1A', value: '' }, // Reserved
-        { type: '10i0', value: 0 }, // Current system pool identifier
-        { type: '10i0', value: 0 }, // Thread count
-        { type: '20u0', value: 0 }, // Processing unit time used - total for the job
-        { type: '20u0', value: 0 }, // Number of auxiliary I/O requests
-        { type: '20u0', value: 0 }, // Processing unit time used for database - total for the job
-        { type: '20u0', value: 0 }, // Page faults
-        { type: '4A', value: '' }, // Active job status for jobs ending
-        { type: '10A', value: '' }, // Memory pool name
-        { type: '1A', value: '' }, // Message reply
-        { type: '4A', value: '' }, // Message key, when active job waiting for a message
-        { type: '10A', value: '' }, // Message queue name, when active job waiting for a message
-        { type: '10A', value: '' }, // Message queue library name, when active job waiting for a message
-        { type: '10A', value: '' }, // Message queue library ASP device name, when active job waiting for a message
-      ],
-    };
-    const JobId = {
-      type: 'ds',
-      fields: [
-        { type: '10A', value: jobName },
-        { type: '10A', value: userName },
-        { type: '6A', value: jobNumber },
-      ],
-    };
-    const pgm = new ProgramCall('QUSRJOBI', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
-    pgm.addParam({ type: '8A', value: 'JOBI0200' });
-    pgm.addParam(JobId);
-    pgm.addParam({ type: '16A', value: '' });
-    pgm.addParam(this.errno);
-
-    this.conn.add(pgm.toXML());
-
-    let rtValue;
-
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-
-          rtValue = {
-            Job_name: data[2]._,
-            User_name: data[3]._,
-            Job_number: data[4]._,
-            Job_status: data[6]._,
-            Job_type: data[7]._,
-            Job_subtype: data[8]._,
-            Subsystem_description_name: data[9]._,
-            'Run_priority_(job)': data[10]._,
-            System_pool_identifier: data[11]._,
-            'Processing_unit_time_used,_if_less_than_2,147,483,647_milliseconds': data[12]._,
-            'Number_of_auxiliary_I/O_requests,_if_less_than_2,147,483,647': data[13]._,
-            Number_of_interactive_transactions: data[14]._,
-            Response_time_total: data[15]._,
-            Function_type: data[16]._,
-            Function_name: data[17]._,
-            Active_job_status: data[18]._,
-            Number_of_database_lock_waits: data[19]._,
-            Number_of_internal_machine_lock_waits: data[20]._,
-            Number_of_nondatabase_lock_waits: data[21]._,
-            Time_spent_on_database_lock_waits: data[22]._,
-            Time_spent_on_internal_machine_lock_waits: data[23]._,
-            Time_spent_on_nondatabase_lock_waits: data[24]._,
-            Current_system_pool_identifier: data[26]._,
-            Thread_count: data[27]._,
-            'Processing_unit_time_used_-_total_for_the_job': data[28]._,
-            'Number_of_auxiliary_I/O_requests': data[29]._,
-            'Processing_unit_time_used_for_database_-_total_for_the_job': data[30]._,
-            Page_faults: data[31]._,
-            Active_job_status_for_jobs_ending: data[32]._,
-            Memory_pool_name: data[33]._,
-            Message_reply: data[34]._,
-            'Message_key,_when_active_job_waiting_for_a_message': data[35]._,
-            'Message_queue_name,_when_active_job_waiting_for_a_message': data[36]._,
-            'Message_queue_library_name,_when_active_job_waiting_for_a_message': data[37]._,
-            'Message_queue_library_ASP_device_name,_when_active_job_waiting_for_a_message': data[38]._,
-          };
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-    this.conn.run(toJson);
-  }
-
-  getDataArea(lib, area, length, cb) {
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      fields: [
-        { type: '10i0', value: 0 }, // Bytes available
-        { type: '10i0', value: 0 }, // Bytes returned
-        { type: '10A', value: '' }, // Type of value returned
-        { type: '10A', value: '' }, // Library name
-        { type: '10i0', value: 0 }, // Length of value returned
-        { type: '10i0', value: 0 }, // Number of decimal positions
-        { type: `${length}A`, value: '' }, // Value
-      ],
-    };
-    const areaPath = {
-      type: 'ds',
-      fields: [
-        { type: '10A', value: area },
-        { type: '10A', value: lib },
-      ],
-    };
-    const pgm = new ProgramCall('QWCRDTAA', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam({ type: '10i0', value: length + 36 });
-    pgm.addParam(areaPath);
-    pgm.addParam({ type: '10i0', value: 1 });
-    pgm.addParam({ type: '10i0', value: length });
-    pgm.addParam(this.errno);
-
-    this.conn.add(pgm.toXML());
-
-    let rtValue;
-
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-          rtValue = {
-            Type_of_value_returned: data[2]._,
-            Library_name: data[3]._,
-            Length_of_value_returned: data[4]._,
-            Number_of_decimal_positions: data[5]._,
-            Value: data[6]._,
-          };
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-    this.conn.run(toJson);
-  }
 }
 
 module.exports.Toolkit = Toolkit;
